File restructure: accounting for how many tests we are going to have with HDF5
