+++ tdms/src/hdf5_io/hdf5_reader.cpp
@@ -1,0 +1,14 @@
+#include "hdf5_io/hdf5_reader.h"
+
+using namespace std;
+
+void HDF5Reader::read(const string &plane, InterfaceComponent *ic) const {
+  // Read the InterfaceComponent in as a 2-element double array
+  double read_buffer[2];
+  read_field_from_struct("interface", plane, read_buffer);
+  // The index that is read in should have 1 subtracted from it, to account for
+  // MATLAB indexing
+  ic->index = max((int) read_buffer[0] - 1, 0);
+  // The apply flag should be cast from the double that is read in
+  ic->apply = (bool) read_buffer[1];
+}
