+++ tdms/tests/unit/hdf5_io_tests/test_hdf5_io.cpp
@@ -2,7 +2,8 @@
  * @file test_hdf5_io.cpp
  * @brief Tests of the HDF5 file I/O functionality.
  */
-#include "hdf5_io.h"
+#include "hdf5_io/hdf5_reader.h"
+#include "hdf5_io/hdf5_writer.h"
 
 // std
 #include <cstdlib>
@@ -23,12 +24,6 @@
 using namespace std;
 using tdms_tests::create_tmp_dir, tdms_tests::uint16s_to_string;
 using tdms_unit_test_data::struct_testdata;
-
-TEST_CASE("Wrong datatype passed to ijk.") {
-  auto tmp = create_tmp_dir();
-  HDF5Writer r(tmp.string() + "/why.h5");
-  // r.data_dump("x");
-}
 
 TEST_CASE("Test file I/O construction/destruction.") {
   // test-case wide setup - temporary directory
@@ -157,58 +152,4 @@ TEST_CASE("Test read/write wrt standard datatypes") {
   // teardown - remove temporary directory and all files
   SPDLOG_DEBUG("Removing temporary directory.");
   filesystem::remove_all(tmp);
-}
-
-TEST_CASE("Read from a MATLAB struct") {
-  HDF5Reader MATFile(struct_testdata);
-
-  SECTION("Read numeric scalars") {
-    /* Read scalar values from the MATLAB struct into the array. */
-    // Initialise with values distinct from the expected values
-    double one_half = 0., unity = 0.;
-    bool logical_read = false;
-    // Read values
-    MATFile.read_field_from_struct("example_struct", "double_half", &one_half);
-    MATFile.read_field_from_struct("example_struct", "double_no_decimal",
-                                   &unity);
-    MATFile.read_field_from_struct("example_struct", "boolean", &logical_read);
-    // Validate read in data
-    REQUIRE(one_half == Catch::Approx(0.5));
-    REQUIRE(unity == Catch::Approx(1.));
-    REQUIRE(int(unity) == 1);
-    REQUIRE(logical_read);
-  }
-
-  SECTION("Read array data") {
-    /* Read in the character array data */
-    // string field is set to "tdms". Note that MATLAB saves this as uint16s, so
-    // we need to convert manually...
-    {
-      uint16_t read_uints16[4];
-      MATFile.read_field_from_struct("example_struct", "string", read_uints16);
-      string tdms = uint16s_to_string(read_uints16, 4);
-      REQUIRE(tdms == "tdms");
-    }
-    // The uint 3*4*5 uint matrix contains only 1s
-    {
-      vector<uint8_t> uint_matrix(3 * 4 * 5, 0);
-      MATFile.read_field_from_struct("example_struct", "uint_345",
-                                     uint_matrix.data());
-      bool all_values_unity = true;
-      for (uint8_t &value : uint_matrix) {
-        if (value != 1) { all_values_unity = false; }
-      }
-      REQUIRE(all_values_unity);
-    }
-    // The double 2*2 matrix contains 0.25, 0.5, 0.75, 1.
-    {
-      double two_by_two[4];
-      MATFile.read_field_from_struct("example_struct", "double_22", two_by_two);
-      REQUIRE(two_by_two[0] == Catch::Approx(0.25));
-      REQUIRE(two_by_two[1] == Catch::Approx(0.75));
-      REQUIRE(two_by_two[2] == Catch::Approx(0.5));
-      REQUIRE(two_by_two[3] == Catch::Approx(1.0));
-    }
-    // The complex matrix is the Pauli-y matrix [0, -i; i, 0]
-  }
 }
